requirements.txt
Django==2.2.2
django-multiselectfield==0.1.8
pytz==2019.1
sqlparse==0.3.0
